Add solargraph and annotate
See also:

- https://github.com/DFE-Digital/apply-for-qualified-teacher-status/pull/19
- https://github.com/DFE-Digital/apply-for-qualified-teacher-status/pull/23

[solargraph](https://github.com/castwide/solargraph) is a Ruby language
server protocol; it provides intellisense capabilities to [compatible
editors](https://github.com/castwide/solargraph#using-solargraph).

`solargraph` can be used without adding a `development` dependency to a
project, and it provides basic Ruby language completion for core
language features. To set it up to work with our bespoke modules, gems,
and Rails, you need to bundle it and run a command. I've documented this
in the README.

The gem is `require: false`d and is not bundled in the production build.

`annotate` and `solargraph-rails` are added for better Rails support.

| Before | After |
| - | - |
| Vim doesn't know what classes are under the ActiveModel::Type namespace | Full class list and inline documentation |
| ![image](https://user-images.githubusercontent.com/1650875/168996614-1569cee6-48aa-455e-ac94-4dd0cc7d5ac4.png) | ![image](https://user-images.githubusercontent.com/1650875/168997430-69c06b0f-fbfb-4466-86f2-07ed94ab6f4d.png) |
| String based autocompletion of instance methods |  Dynamic introspection |
| ![image](https://user-images.githubusercontent.com/1650875/168996796-495d6fb4-b0fb-4240-896b-eae0aad66943.png) | ![image](https://user-images.githubusercontent.com/1650875/168997511-44188867-e089-495c-a61a-a7b76d05de73.png) |