ensure asdf gets the right version of postgres
When bundle runs, it leaves the directory that has .tools-versions in
it, which means asdf loses track of which version of postgres we want
and building the native components of the "pg" gem fail.

By setting the env var we ensure asdf picks out the right version of
postgres when we run bundle.